Revert "Merge branch 'milestone-2' into MAR-48-Vendor-Inventory"
This reverts commit b135df29e9cdd69b869d097c3880b76b5fa6d347, reversing
changes made to 81203f973e246502c99aaffab22f1db3311f1d5e.

+++ backend/src/main/java/com/market/backend/models/Comment.java
@@ -27,5 +27,5 @@ public class Comment {
     private String body;
 
     @Column(name = "rating", nullable = false)
-    private Integer rating;
+    private double rating;
 }
